1. TransparentBlt 함수로 플레이어 이미지 렌더링 (마젠타로 투명처리)
2. Task Manager 구현 중

+++ GameClient/qPathMgr.cpp
@@ -48,7 +48,7 @@ void qPathMgr::GetParentPath(_Inout_ wchar_t* _Buffer)
 {
 	size_t len = wcslen(_Buffer);
 
-	for (int i = len - 1; 0 < i; --i)
+	for (size_t i = len - 1; 0 < i; --i)
 	{
 		if (L'\\' == _Buffer[i])
 		{
